requirements.txt
fastapi
uvicorn
ollama
langgraph
gitpython
requests
pytest
pyyaml
python-slugify
typing-extensions
